refactor: exec types
