+++ lib/util/exec/index.ts
@@ -1,30 +1,20 @@
-import type { ExecOptions as ChildProcessExecOptions } from 'child_process';
 import { dirname, join } from 'upath';
 import { GlobalConfig } from '../../config/global';
 import { TEMPORARY_ERROR } from '../../constants/error-messages';
 import { logger } from '../../logger';
 import { generateInstallCommands } from './buildpack';
-import {
-  DockerExtraCommands,
-  DockerOptions,
-  ExecResult,
-  Opt,
-  RawExecOptions,
-  rawExec,
-} from './common';
+import { rawExec } from './common';
 import { generateDockerCommand, removeDockerContainer } from './docker';
 import { getChildProcessEnv } from './env';
-import type { ToolConstraint } from './types';
+import type {
+  DockerOptions,
+  ExecOptions,
+  ExecResult,
+  ExtraEnv,
+  RawExecOptions,
+} from './types';
 
-type ExtraEnv<T = unknown> = Record<string, T>;
-
-export interface ExecOptions extends ChildProcessExecOptions {
-  cwdFile?: string;
-  extraEnv?: Opt<ExtraEnv>;
-  docker?: Opt<DockerOptions>;
-  toolConstraints?: Opt<ToolConstraint[]>;
-  preCommands?: DockerExtraCommands;
-}
+export * from './types';
 
 function getChildEnv({
   extraEnv = {},
@@ -71,20 +61,14 @@ function getCwd({ cwd, cwdFile }: ExecOptions): string {
 
 function getRawExecOptions(opts: ExecOptions): RawExecOptions {
   const defaultExecutionTimeout = GlobalConfig.get('executionTimeout');
-  const execOptions: ExecOptions = { ...opts };
-  delete execOptions.extraEnv;
-  delete execOptions.docker;
-  delete execOptions.cwdFile;
-  delete execOptions.toolConstraints;
-  delete execOptions.preCommands;
-
   const childEnv = getChildEnv(opts);
   const cwd = getCwd(opts);
   const rawExecOptions: RawExecOptions = {
-    encoding: 'utf-8',
-    ...execOptions,
-    env: childEnv,
     cwd,
+    encoding: 'utf-8',
+    env: childEnv,
+    maxBuffer: opts.maxBuffer,
+    timeout: opts.timeout,
   };
   // Set default timeout config.executionTimeout if specified; othrwise to 15 minutes
   if (!rawExecOptions.timeout) {
